Final Assignment
Hi,
I was unable to scrape the tables from the webpage.i was however, able to pull the cities.

It is also why i dont have the code for the ML part up. I understand if you give me a zero, but I would appreciate if you could send me the code or atleast a part ofit, so i could understand how to doit.

Thanks again for a great course,
Anirudh